+++ tests/Forge.Tests/Tests.fs
@@ -87,6 +87,18 @@ let tests =
         Expect.equal s.AssemblyName.Data (Some "Test") "should be same"
         Expect.equal s.DocumentationFile.Data (Some "bin\Debug\Test.XML") "should be same"
 
+      testCase "parse - gets all multi target frameworks" <| fun _ ->
+        let projectFile = FsProject.parse netCoreProjectMultiTargetsNoFiles
+        let s = projectFile.Settings
+        Expect.equal s.TargetFrameworks.Data (Some ["net461"; "netstandard2.0"; "netcoreapp2.0"]) "should be same"
+
+      testCase "parse - ToXElem sets all multi target frameworks" <| fun _ ->
+        let projectFile = FsProject.parse netCoreProjectMultiTargetsNoFiles
+        let s = projectFile.Settings
+        let settingsXml = projectFile.Settings.ToXElem()
+        let targetFrameworks = settingsXml.Element (Xml.Linq.XName.Get "TargetFrameworks")
+        Expect.equal targetFrameworks.Value "net461;netstandard2.0;netcoreapp2.0" "should be same"
+
       testCase "parse - add new file" <| fun _ ->
         let pf = FsProject.parse astInput
         let f = {SourceFile.Include = "Test.fsi"; Condition = None; OnBuild = BuildAction.Compile; Link = None; Copy = None; Paket = None}
@@ -144,6 +156,22 @@ let tests =
         Expect.equal s.RootNamespace.Data (Some "TestRename") "should be same"
         Expect.equal s.DocumentationFile.Data (Some "bin\Debug\TestRename.XML") "should be same"
 
+      testCase "parse - rename file" <| fun _ ->
+        let pf = FsProject.parse astInput
+        let pf' = pf |> FsProject.renameFile "FixProject.fs" "renamed_file.fs"
+        let files = pf'.SourceFiles.AllFiles()
+        Expect.equal (files |> Seq.head) "renamed_file.fs" "should be same"
+        files |> Expect.hasLength 3
+
+      testCase "parse - rename file invalid name" <| fun _ ->
+        if System.IO.Path.GetInvalidFileNameChars().Length > 0 then
+            let invalid = System.IO.Path.GetInvalidFileNameChars().[0].ToString()
+            let pf = FsProject.parse astInput
+            let pf' = pf |> FsProject.renameFile "FixProject.fs" ("invalid" + invalid + ".fs")
+            let files = pf'.SourceFiles.AllFiles()
+            Expect.equal (files |> Seq.head) "FixProject.fs" "should be same"
+            files |> Expect.hasLength 3
+
       testCase "parse - move up" <| fun _ ->
         let pf = FsProject.parse astInput
         let pf' = pf |> FsProject.moveUp "a_file.fs"
@@ -157,6 +185,22 @@ let tests =
         let files = pf'.SourceFiles.AllFiles()
         Expect.equal (files |> Seq.head) "App.config" "should be same"
         files |> Expect.hasLength 3
+
+      testCase "parse - add above" <| fun _ ->
+        let pf = FsProject.parse astInput
+        let f = {SourceFile.Include = "above.fs"; Condition = None; OnBuild = BuildAction.Compile; Link = None; Copy = None; Paket = None}
+        let pf' = FsProject.addAbove "FixProject.fs" f pf
+        let files = pf'.SourceFiles.AllFiles()
+        Expect.equal (files |> Seq.head) "above.fs" "should be same"
+        pf'.SourceFiles.AllFiles() |> Expect.hasLength 4
+
+      testCase "parse - add below" <| fun _ ->
+        let pf = FsProject.parse astInput
+        let f = {SourceFile.Include = "below.fs"; Condition = None; OnBuild = BuildAction.Compile; Link = None; Copy = None; Paket = None}
+        let pf' = FsProject.addBelow "FixProject.fs" f pf
+        let files = pf'.SourceFiles.AllFiles()
+        Expect.equal (files |> Seq.item 1) "below.fs" "should be same"
+        pf'.SourceFiles.AllFiles() |> Expect.hasLength 4
     ]
 
     testList "SolutionSystem" [